+++ api/src/main/java/org/openmrs/module/episodes/service/impl/EpisodeAttributeTypeServiceImpl.java
@@ -2,7 +2,7 @@ package org.openmrs.module.episodes.service.impl;
 
 import org.openmrs.api.context.Context;
 import org.openmrs.module.episodes.EpisodeAttributeType;
-import org.openmrs.module.episodes.dao.impl.EpisodeAttributeTypeDao;
+import org.openmrs.module.episodes.dao.EpisodeAttributeTypeDao;
 import org.openmrs.module.episodes.service.EpisodeAttributeTypeService;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
